tests: reorganise into tests subdirectory
The tests have been moved into a tests directory and split into
functional groups.
