+++ tests/test_get_xpath.py
@@ -1,0 +1,229 @@
+import pytest
+from conftest import _get_test_with_filter
+
+
+# GET XPATH
+
+
+def test_get_xpath_node():
+    xpath = '/test/settings/debug'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <settings>
+            <debug>enable</debug>
+        </settings>
+    </test>
+</nc:data>
+    """
+    xml = _get_test_with_filter(xpath, expected, f_type='xpath')
+    assert xml.find('./{*}test/{*}settings/{*}debug').text == 'enable'
+
+
+# No default or prefixed namespace - we use the internal default namespace
+def test_get_xpath_node_ns_none():
+    xpath = '/test/settings/priority'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <settings>
+            <priority>1</priority>
+        </settings>
+    </test>
+</nc:data>
+    """
+    xml = _get_test_with_filter(xpath, expected, f_type='xpath')
+    assert xml.find('./{*}test/{*}settings/{*}priority').text == '1'
+
+
+def test_get_xpath_node_ns_aug_none():
+    xpath = '/test/settings/volume'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <settings>
+            <volume>1</volume>
+        </settings>
+    </test>
+</nc:data>
+    """
+    xml = _get_test_with_filter(xpath, expected, f_type='xpath')
+    assert xml.find('./{*}test/{*}settings/{*}volume').text == '1'
+
+
+# Default namespace is the same as the internal default namespace
+@pytest.mark.skip(reason="does not work yet")
+def test_get_xpath_node_ns_default():
+    xpath = '/test:test/settings/priority'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <settings>
+            <priority>1</priority>
+        </settings>
+    </test>
+</nc:data>
+    """
+    xml = _get_test_with_filter(xpath, expected, f_type='xpath')
+    assert xml.find('./{*}test/{*}settings/{*}priority').text == '1'
+
+
+@pytest.mark.skip(reason="does not work yet")
+def test_get_xpath_node_ns_aug_default():
+    xpath = '/test:test/settings/volume'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <settings>
+            <volume>1</volume>
+        </settings>
+    </test>
+</nc:data>
+    """
+    xml = _get_test_with_filter(xpath, expected, f_type='xpath')
+    assert xml.find('./{*}test/{*}settings/{*}volume').text == '1'
+
+
+# Prefixed namespace is not the internal default namespace
+def test_get_xpath_node_ns_other():
+    xpath = '/t2:test/t2:settings/t2:priority'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <settings>
+            <priority>2</priority>
+        </settings>
+    </test>
+</nc:data>
+    """
+    xml = _get_test_with_filter(xpath, expected, f_type='xpath')
+    assert xml.find('./{*}test/{*}settings/{*}priority').text == '2'
+
+
+def test_get_xpath_node_ns_aug_other():
+    xpath = '/t2:test/t2:settings/aug2:speed'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <settings>
+            <speed>2</speed>
+        </settings>
+    </test>
+</nc:data>
+    """
+    xml = _get_test_with_filter(xpath, expected, f_type='xpath')
+    assert xml.find('./{*}test/{*}settings/{*}speed').text == '2'
+
+
+def test_get_xpath_trunk():
+    xpath = '/test/settings'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <settings>
+            <debug>enable</debug>
+            <enable>true</enable>
+            <priority>1</priority>
+            <volume>1</volume>
+        </settings>
+    </test>
+</nc:data>
+    """
+    _get_test_with_filter(xpath, expected, f_type='xpath')
+
+
+def test_get_xpath_list_trunk():
+    xpath = '/test/animals'
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+  <test>
+    <animals>
+      <animal>
+        <name>cat</name>
+        <type>big</type>
+      </animal>
+      <animal>
+        <name>dog</name>
+        <colour>brown</colour>
+      </animal>
+      <animal>
+        <name>hamster</name>
+        <type>little</type>
+        <food>
+          <name>banana</name>
+          <type>fruit</type>
+        </food>
+        <food>
+          <name>nuts</name>
+          <type>kibble</type>
+        </food>
+      </animal>
+      <animal>
+        <name>mouse</name>
+        <type>little</type>
+        <colour>grey</colour>
+      </animal>
+      <animal>
+        <name>parrot</name>
+        <type>big</type>
+        <colour>blue</colour>
+      </animal>
+    </animals>
+  </test>
+</nc:data>
+    """
+    _get_test_with_filter(xpath, expected, f_type='xpath')
+
+
+def test_get_xpath_list_select_one_trunk():
+    xpath = "/test/animals/animal[name='cat']"
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <animals>
+            <animal>
+                <name>cat</name>
+                <type>big</type>
+            </animal>
+        </animals>
+    </test>
+</nc:data>
+    """
+    _get_test_with_filter(xpath, expected, f_type='xpath')
+
+
+def test_get_xpath_list_select_one_parameter():
+    xpath = "/test/animals/animal[name='cat']/type"
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <animals>
+            <animal>
+                <name>cat</name>
+                <type>big</type>
+            </animal>
+        </animals>
+    </test>
+</nc:data>
+    """
+    _get_test_with_filter(xpath, expected, f_type='xpath')
+
+
+@pytest.mark.skip(reason="does not work yet")
+def test_xpath_query_multi():
+    xpath = ("/test/animals/animal[name='cat']/name | /test/animals/animal[name='dog']/name")
+    expected = """
+<nc:data xmlns:nc="urn:ietf:params:xml:ns:netconf:base:1.0">
+    <test>
+        <animals>
+            <animal>
+                <name>cat</name>
+            </animal>
+            <animal>
+                <name>dog</name>
+            </animal>
+        </animals>
+    </test>
+</nc:data>
+    """
+    _get_test_with_filter(xpath, expected, f_type='xpath')
